requirements.txt
py-cord
ezcord
python-dotenv
pymongo